requirements.txt
# Core audio processing
sounddevice>=0.4.6
numpy>=1.21.0
scipy>=1.7.0

# GUI and visualization
PyQt6>=6.4.0
matplotlib>=3.5.0

# Configuration and utilities
pyyaml>=6.0
argparse>=1.4.0